Split mutation reducer lifecycle handlers
