+++ tests/test_state_reducer.py
@@ -5,77 +5,31 @@ from tests.state_test_helpers import reduce_state
 from zivo.models import (
     AppConfig,
     BookmarkConfig,
-    CreatePathRequest,
-    CreateZipArchiveRequest,
-    CreateZipArchiveResult,
-    DeleteRequest,
     ExternalLaunchRequest,
-    ExtractArchiveRequest,
-    ExtractArchiveResult,
-    FileMutationResult,
-    PasteAppliedChange,
-    PasteConflict,
-    PasteRequest,
-    PasteSummary,
-    RenameRequest,
-    TrashRestoreRecord,
-    UndoDeletePathStep,
-    UndoEntry,
-    UndoMovePathStep,
-    UndoRestoreTrashStep,
-    UndoResult,
 )
 from zivo.state import (
     ActivateNextTab,
     ActivatePreviousTab,
     AddBookmark,
-    ArchiveExtractCompleted,
-    ArchiveExtractConfirmationState,
-    ArchiveExtractFailed,
-    ArchiveExtractProgress,
-    ArchiveExtractProgressState,
-    ArchivePreparationCompleted,
-    ArchivePreparationFailed,
-    BeginCreateInput,
-    BeginDeleteTargets,
-    BeginEmptyTrash,
-    BeginExtractArchiveInput,
     BeginFilterInput,
     BeginHistorySearch,
-    BeginRenameInput,
-    BeginZipCompressInput,
     BrowserSnapshot,
     BrowserSnapshotFailed,
     BrowserSnapshotLoaded,
-    CancelArchiveExtractConfirmation,
-    CancelDeleteConfirmation,
-    CancelEmptyTrashConfirmation,
     CancelFilterInput,
-    CancelPasteConflict,
-    CancelPendingInput,
-    CancelZipCompressConfirmation,
     ChildPaneSnapshotFailed,
     ChildPaneSnapshotLoaded,
     ClearSelection,
-    ClipboardPasteCompleted,
-    ClipboardPasteFailed,
-    ClipboardPasteNeedsResolution,
     CloseCurrentTab,
     CloseSplitTerminalEffect,
     ConfigEditorState,
     ConfigSaveCompleted,
     ConfigSaveFailed,
-    ConfirmArchiveExtract,
-    ConfirmDeleteTargets,
-    ConfirmEmptyTrash,
     ConfirmFilterInput,
-    ConfirmZipCompress,
     CopyPathsToClipboard,
-    CopyTargets,
     CurrentPaneDeltaState,
     CutTargets,
     CycleConfigEditorValue,
-    DeleteConfirmationState,
     DirectoryEntryState,
     DirectorySizeCacheEntry,
     DirectorySizeDeltaState,
@@ -83,12 +37,9 @@ from zivo.state import (
     DirectorySizesLoaded,
     DismissConfigEditor,
     DismissNameConflict,
-    EmptyTrashConfirmationState,
     EnterCursorDirectory,
     ExternalLaunchCompleted,
     ExternalLaunchFailed,
-    FileMutationCompleted,
-    FileMutationFailed,
     FocusSplitTerminal,
     GoBack,
     GoForward,
@@ -109,32 +60,20 @@ from zivo.state import (
     OpenPathWithDefaultApp,
     OpenTerminalAtPath,
     PaneState,
-    PasteClipboard,
-    PasteConflictState,
     PendingInputState,
     PendingKeySequenceState,
     ReloadDirectory,
     RemoveBookmark,
     RequestBrowserSnapshot,
     RequestDirectorySizes,
-    ResolvePasteConflict,
-    RunArchiveExtractEffect,
-    RunArchivePreparationEffect,
-    RunClipboardPasteEffect,
     RunConfigSaveEffect,
     RunDirectorySizeEffect,
     RunExternalLaunchEffect,
-    RunFileMutationEffect,
-    RunUndoEffect,
-    RunZipCompressEffect,
-    RunZipCompressPreparationEffect,
     SaveConfigEditor,
-    SelectAllVisibleEntries,
     SendSplitTerminalInput,
     SetCursorPath,
     SetFilterQuery,
     SetNotification,
-    SetPendingInputValue,
     SetPendingKeySequence,
     SetSort,
     SetTerminalHeight,
@@ -143,26 +82,14 @@ from zivo.state import (
     SplitTerminalOutputReceived,
     SplitTerminalStarted,
     StartSplitTerminalEffect,
-    SubmitPendingInput,
     ToggleHiddenFiles,
     ToggleSelection,
-    ToggleSelectionAndAdvance,
     ToggleSplitTerminal,
-    UndoCompleted,
-    UndoFailed,
-    UndoLastOperation,
     WriteSplitTerminalInputEffect,
-    ZipCompressCompleted,
-    ZipCompressConfirmationState,
-    ZipCompressFailed,
-    ZipCompressPreparationCompleted,
-    ZipCompressProgress,
-    ZipCompressProgressState,
     build_initial_app_state,
     reduce_app_state,
     select_browser_tabs,
 )
-from zivo.state.reducer_common import browser_snapshot_invalidation_paths
 
 
 def _reduce_state(state, action):
@@ -812,70 +739,6 @@ def test_begin_filter_input_switches_mode_without_mutating_query() -> None:
     assert next_state.ui_mode == "FILTER"
     assert next_state.filter == state.filter
 
-def test_begin_rename_input_sets_initial_value_from_target_name() -> None:
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(state, BeginRenameInput("/home/tadashi/develop/zivo/docs"))
-
-    assert next_state.ui_mode == "RENAME"
-    assert next_state.pending_input == PendingInputState(
-        prompt="Rename: ",
-        value="docs",
-        cursor_pos=4,
-        target_path="/home/tadashi/develop/zivo/docs",
-    )
-
-def test_begin_rename_input_ignores_unknown_path() -> None:
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(state, BeginRenameInput("/tmp/missing"))
-
-    assert next_state == state
-
-def test_begin_create_input_sets_mode_and_kind() -> None:
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(state, BeginCreateInput("dir"))
-
-    assert next_state.ui_mode == "CREATE"
-    assert next_state.pending_input == PendingInputState(
-        prompt="New directory: ",
-        value="",
-        create_kind="dir",
-    )
-
-def test_begin_extract_archive_input_sets_default_destination() -> None:
-    archive_path = "/home/tadashi/develop/zivo/archive.tar.gz"
-    expected_value = str(Path("/home/tadashi/develop/zivo/archive").resolve())
-    next_state = _reduce_state(
-        build_initial_app_state(),
-        BeginExtractArchiveInput(archive_path),
-    )
-
-    assert next_state.ui_mode == "EXTRACT"
-    assert next_state.pending_input == PendingInputState(
-        prompt="Extract to: ",
-        value=expected_value,
-        cursor_pos=len(expected_value),
-        extract_source_path=archive_path,
-    )
-
-def test_begin_zip_compress_input_sets_default_destination() -> None:
-    source_path = "/home/tadashi/develop/zivo/README.md"
-    expected_value = str(Path("/home/tadashi/develop/zivo/README.zip").resolve())
-    next_state = _reduce_state(
-        build_initial_app_state(),
-        BeginZipCompressInput((source_path,)),
-    )
-
-    assert next_state.ui_mode == "ZIP"
-    assert next_state.pending_input == PendingInputState(
-        prompt="Compress to: ",
-        value=expected_value,
-        cursor_pos=len(expected_value),
-        zip_source_paths=(source_path,),
-    )
-
 def test_move_config_editor_cursor_clamps_to_visible_settings() -> None:
     state = replace(
         build_initial_app_state(config_path="/tmp/zivo/config.toml"),
@@ -1467,49 +1330,6 @@ def test_split_terminal_exited_resets_state_and_notifies() -> None:
         message="Split terminal closed (exit 0)",
     )
 
-def test_select_all_visible_entries_replaces_selection_with_visible_paths() -> None:
-    initial_state = build_initial_app_state()
-    state = replace(
-        initial_state,
-        current_pane=PaneState(
-            directory_path="/home/tadashi/develop/zivo",
-            entries=(
-                DirectoryEntryState(
-                    "/home/tadashi/develop/zivo/.env",
-                    ".env",
-                    "file",
-                    hidden=True,
-                ),
-                DirectoryEntryState("/home/tadashi/develop/zivo/docs", "docs", "dir"),
-                DirectoryEntryState("/home/tadashi/develop/zivo/src", "src", "dir"),
-            ),
-            cursor_path="/home/tadashi/develop/zivo/docs",
-            selected_paths=frozenset(
-                {
-                    "/home/tadashi/develop/zivo/.env",
-                    "/home/tadashi/develop/zivo/docs",
-                }
-            ),
-        ),
-    )
-
-    next_state = _reduce_state(
-        state,
-        SelectAllVisibleEntries(
-            (
-                "/home/tadashi/develop/zivo/docs",
-                "/home/tadashi/develop/zivo/src",
-            )
-        ),
-    )
-
-    assert next_state.current_pane.selected_paths == frozenset(
-        {
-            "/home/tadashi/develop/zivo/docs",
-            "/home/tadashi/develop/zivo/src",
-        }
-    )
-
 def test_toggle_hidden_files_normalizes_cursor_and_selection() -> None:
     hidden_path = "/home/tadashi/develop/zivo/.env"
     visible_path = "/home/tadashi/develop/zivo/docs"
@@ -1539,795 +1359,6 @@ def test_toggle_hidden_files_normalizes_cursor_and_selection() -> None:
         message="Hidden files hidden",
     )
 
-def test_begin_delete_targets_single_runs_file_mutation() -> None:
-    state = build_initial_app_state(confirm_delete=False)
-
-    result = reduce_app_state(
-        state,
-        BeginDeleteTargets(("/home/tadashi/develop/zivo/docs",)),
-    )
-
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects == (
-        RunFileMutationEffect(
-            request_id=1,
-            request=DeleteRequest(paths=("/home/tadashi/develop/zivo/docs",), mode="trash"),
-        ),
-    )
-
-def test_begin_delete_targets_single_enters_confirm_mode_when_enabled() -> None:
-    state = build_initial_app_state(confirm_delete=True)
-
-    next_state = _reduce_state(
-        state,
-        BeginDeleteTargets(("/home/tadashi/develop/zivo/docs",)),
-    )
-
-    assert next_state.ui_mode == "CONFIRM"
-    assert next_state.delete_confirmation == DeleteConfirmationState(
-        paths=("/home/tadashi/develop/zivo/docs",)
-    )
-
-def test_begin_delete_targets_with_empty_paths_keeps_state() -> None:
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(state, BeginDeleteTargets(()))
-
-    assert next_state == state
-
-def test_begin_delete_targets_multiple_enters_confirm_mode() -> None:
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(
-        state,
-        BeginDeleteTargets(
-            (
-                "/home/tadashi/develop/zivo/docs",
-                "/home/tadashi/develop/zivo/src",
-            )
-        ),
-    )
-
-    assert next_state.ui_mode == "CONFIRM"
-    assert next_state.delete_confirmation == DeleteConfirmationState(
-        paths=(
-            "/home/tadashi/develop/zivo/docs",
-            "/home/tadashi/develop/zivo/src",
-        )
-    )
-
-def test_cancel_pending_input_returns_to_browsing() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="RENAME",
-        pending_input=PendingInputState(prompt="Rename: ", value="docs"),
-    )
-
-    next_state = _reduce_state(state, CancelPendingInput())
-
-    assert next_state.ui_mode == "BROWSING"
-    assert next_state.pending_input is None
-
-def test_confirm_delete_targets_runs_file_mutation() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        delete_confirmation=DeleteConfirmationState(
-            paths=(
-                "/home/tadashi/develop/zivo/docs",
-                "/home/tadashi/develop/zivo/src",
-            )
-        ),
-        next_request_id=4,
-    )
-
-    result = reduce_app_state(state, ConfirmDeleteTargets())
-
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects == (
-        RunFileMutationEffect(
-            request_id=4,
-            request=DeleteRequest(
-                paths=(
-                    "/home/tadashi/develop/zivo/docs",
-                    "/home/tadashi/develop/zivo/src",
-                ),
-                mode="trash",
-            ),
-        ),
-    )
-
-def test_cancel_delete_confirmation_returns_to_browsing_with_warning() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        delete_confirmation=DeleteConfirmationState(
-            paths=("/home/tadashi/develop/zivo/docs",),
-        ),
-    )
-
-    next_state = _reduce_state(state, CancelDeleteConfirmation())
-
-    assert next_state.ui_mode == "BROWSING"
-    assert next_state.notification == NotificationState(level="warning", message="Delete cancelled")
-
-def test_begin_permanent_delete_targets_enters_confirm_mode_when_delete_confirmation_disabled(
-) -> None:
-    state = build_initial_app_state(confirm_delete=False)
-
-    next_state = _reduce_state(
-        state,
-        BeginDeleteTargets(("/home/tadashi/develop/zivo/docs",), mode="permanent"),
-    )
-
-    assert next_state.ui_mode == "CONFIRM"
-    assert next_state.delete_confirmation == DeleteConfirmationState(
-        paths=("/home/tadashi/develop/zivo/docs",),
-        mode="permanent",
-    )
-
-def test_confirm_permanent_delete_targets_runs_file_mutation() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        delete_confirmation=DeleteConfirmationState(
-            paths=("/home/tadashi/develop/zivo/docs",),
-            mode="permanent",
-        ),
-        next_request_id=4,
-    )
-
-    result = reduce_app_state(state, ConfirmDeleteTargets())
-
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects == (
-        RunFileMutationEffect(
-            request_id=4,
-            request=DeleteRequest(
-                paths=("/home/tadashi/develop/zivo/docs",),
-                mode="permanent",
-            ),
-        ),
-    )
-
-def test_cancel_permanent_delete_confirmation_returns_to_browsing_with_warning() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        delete_confirmation=DeleteConfirmationState(
-            paths=("/home/tadashi/develop/zivo/docs",),
-            mode="permanent",
-        ),
-    )
-
-    next_state = _reduce_state(state, CancelDeleteConfirmation())
-
-    assert next_state.ui_mode == "BROWSING"
-    assert next_state.notification == NotificationState(
-        level="warning",
-        message="Permanent delete cancelled",
-    )
-
-def test_submit_pending_extract_starts_archive_preparation() -> None:
-    source_path = "/home/tadashi/develop/zivo/archive.zip"
-    dest_path = "/tmp/output/archive"
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="EXTRACT",
-        pending_input=PendingInputState(
-            prompt="Extract to: ",
-            value=dest_path,
-            extract_source_path=source_path,
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.state.pending_archive_prepare_request_id == 1
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects == (
-        RunArchivePreparationEffect(
-            request_id=1,
-            request=ExtractArchiveRequest(
-                source_path=str(Path(source_path).resolve()),
-                destination_path=str(Path(dest_path).resolve()),
-            ),
-        ),
-    )
-
-def test_submit_pending_zip_compress_starts_preparation() -> None:
-    dest_path = "/tmp/output.zip"
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="ZIP",
-        pending_input=PendingInputState(
-            prompt="Compress to: ",
-            value=dest_path,
-            zip_source_paths=(
-                "/home/tadashi/develop/zivo/docs",
-                "/home/tadashi/develop/zivo/src",
-            ),
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.state.pending_zip_compress_prepare_request_id == 1
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects == (
-        RunZipCompressPreparationEffect(
-            request_id=1,
-            request=CreateZipArchiveRequest(
-                source_paths=(
-                    "/home/tadashi/develop/zivo/docs",
-                    "/home/tadashi/develop/zivo/src",
-                ),
-                destination_path=str(Path(dest_path).resolve()),
-                root_dir="/home/tadashi/develop/zivo",
-            ),
-        ),
-    )
-
-def test_submit_pending_extract_resolves_relative_destination_from_archive_parent() -> None:
-    source_path = "/home/tadashi/develop/zivo/docs/archive.tar.bz2"
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="EXTRACT",
-        pending_input=PendingInputState(
-            prompt="Extract to: ",
-            value="../exports/archive",
-            extract_source_path=source_path,
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.effects == (
-        RunArchivePreparationEffect(
-            request_id=1,
-            request=ExtractArchiveRequest(
-                source_path=str(Path(source_path).resolve()),
-                destination_path=str(Path("/home/tadashi/develop/zivo/exports/archive").resolve()),
-            ),
-        ),
-    )
-
-def test_archive_preparation_with_conflicts_enters_confirm_mode() -> None:
-    request = ExtractArchiveRequest(
-        source_path="/home/tadashi/develop/zivo/archive.zip",
-        destination_path="/tmp/output/archive",
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_input=PendingInputState(
-            prompt="Extract to: ",
-            value="/tmp/output/archive",
-            extract_source_path=request.source_path,
-        ),
-        pending_archive_prepare_request_id=4,
-    )
-
-    next_state = _reduce_state(
-        state,
-        ArchivePreparationCompleted(
-            request_id=4,
-            request=request,
-            total_entries=7,
-            conflict_count=2,
-            first_conflict_path="/tmp/output/archive/notes.txt",
-        ),
-    )
-
-    assert next_state.ui_mode == "CONFIRM"
-    assert next_state.pending_archive_prepare_request_id is None
-    assert next_state.archive_extract_confirmation == ArchiveExtractConfirmationState(
-        request=request,
-        conflict_count=2,
-        first_conflict_path="/tmp/output/archive/notes.txt",
-        total_entries=7,
-    )
-
-def test_zip_compress_preparation_with_existing_destination_enters_confirm_mode() -> None:
-    request = CreateZipArchiveRequest(
-        source_paths=("/home/tadashi/develop/zivo/docs",),
-        destination_path="/home/tadashi/develop/zivo/docs.zip",
-        root_dir="/home/tadashi/develop/zivo",
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_input=PendingInputState(
-            prompt="Compress to: ",
-            value="/home/tadashi/develop/zivo/docs.zip",
-            zip_source_paths=request.source_paths,
-        ),
-        pending_zip_compress_prepare_request_id=4,
-    )
-
-    next_state = _reduce_state(
-        state,
-        ZipCompressPreparationCompleted(
-            request_id=4,
-            request=request,
-            total_entries=7,
-            destination_exists=True,
-        ),
-    )
-
-    assert next_state.ui_mode == "CONFIRM"
-    assert next_state.pending_zip_compress_prepare_request_id is None
-    assert next_state.zip_compress_confirmation == ZipCompressConfirmationState(
-        request=request,
-        total_entries=7,
-    )
-
-def test_confirm_archive_extract_runs_extract_effect() -> None:
-    request = ExtractArchiveRequest(
-        source_path="/home/tadashi/develop/zivo/archive.zip",
-        destination_path="/tmp/output/archive",
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        pending_input=PendingInputState(
-            prompt="Extract to: ",
-            value="/tmp/output/archive",
-            extract_source_path=request.source_path,
-        ),
-        archive_extract_confirmation=ArchiveExtractConfirmationState(
-            request=request,
-            conflict_count=1,
-            first_conflict_path="/tmp/output/archive/notes.txt",
-            total_entries=3,
-        ),
-    )
-
-    result = reduce_app_state(state, ConfirmArchiveExtract())
-
-    assert result.state.pending_archive_extract_request_id == 1
-    assert result.effects == (
-        RunArchiveExtractEffect(
-            request_id=1,
-            request=request,
-        ),
-    )
-
-def test_confirm_zip_compress_runs_effect() -> None:
-    request = CreateZipArchiveRequest(
-        source_paths=("/home/tadashi/develop/zivo/docs",),
-        destination_path="/home/tadashi/develop/zivo/docs.zip",
-        root_dir="/home/tadashi/develop/zivo",
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        pending_input=PendingInputState(
-            prompt="Compress to: ",
-            value="/home/tadashi/develop/zivo/docs.zip",
-            zip_source_paths=request.source_paths,
-        ),
-        zip_compress_confirmation=ZipCompressConfirmationState(
-            request=request,
-            total_entries=3,
-        ),
-    )
-
-    result = reduce_app_state(state, ConfirmZipCompress())
-
-    assert result.state.pending_zip_compress_request_id == 1
-    assert result.effects == (
-        RunZipCompressEffect(
-            request_id=1,
-            request=request,
-        ),
-    )
-
-def test_cancel_archive_extract_confirmation_returns_to_extract_mode() -> None:
-    request = ExtractArchiveRequest(
-        source_path="/home/tadashi/develop/zivo/archive.zip",
-        destination_path="/tmp/output/archive",
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        pending_input=PendingInputState(
-            prompt="Extract to: ",
-            value="/tmp/output/archive",
-            extract_source_path=request.source_path,
-        ),
-        archive_extract_confirmation=ArchiveExtractConfirmationState(
-            request=request,
-            conflict_count=1,
-            first_conflict_path="/tmp/output/archive/notes.txt",
-            total_entries=3,
-        ),
-    )
-
-    next_state = _reduce_state(state, CancelArchiveExtractConfirmation())
-
-    assert next_state.ui_mode == "EXTRACT"
-    assert next_state.archive_extract_confirmation is None
-    assert next_state.notification == NotificationState(
-        level="warning",
-        message="Extraction cancelled",
-    )
-
-def test_cancel_zip_compress_confirmation_returns_to_zip_mode() -> None:
-    request = CreateZipArchiveRequest(
-        source_paths=("/home/tadashi/develop/zivo/docs",),
-        destination_path="/home/tadashi/develop/zivo/docs.zip",
-        root_dir="/home/tadashi/develop/zivo",
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        pending_input=PendingInputState(
-            prompt="Compress to: ",
-            value="/home/tadashi/develop/zivo/docs.zip",
-            zip_source_paths=request.source_paths,
-        ),
-        zip_compress_confirmation=ZipCompressConfirmationState(
-            request=request,
-            total_entries=3,
-        ),
-    )
-
-    next_state = _reduce_state(state, CancelZipCompressConfirmation())
-
-    assert next_state.ui_mode == "ZIP"
-    assert next_state.zip_compress_confirmation is None
-    assert next_state.notification == NotificationState(
-        level="warning",
-        message="Zip compression cancelled",
-    )
-
-def test_archive_extract_progress_updates_notification() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_archive_extract_request_id=6,
-    )
-
-    next_state = _reduce_state(
-        state,
-        ArchiveExtractProgress(
-            request_id=6,
-            completed_entries=2,
-            total_entries=5,
-            current_path="/tmp/output/archive/notes.txt",
-        ),
-    )
-
-    assert next_state.archive_extract_progress == ArchiveExtractProgressState(
-        completed_entries=2,
-        total_entries=5,
-        current_path="/tmp/output/archive/notes.txt",
-    )
-    assert next_state.notification == NotificationState(
-        level="info",
-        message="Extracting archive 2/5: notes.txt",
-    )
-
-def test_zip_compress_progress_updates_notification() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_zip_compress_request_id=6,
-    )
-
-    next_state = _reduce_state(
-        state,
-        ZipCompressProgress(
-            request_id=6,
-            completed_entries=2,
-            total_entries=5,
-            current_path="/home/tadashi/develop/zivo/docs/readme.txt",
-        ),
-    )
-
-    assert next_state.zip_compress_progress == ZipCompressProgressState(
-        completed_entries=2,
-        total_entries=5,
-        current_path="/home/tadashi/develop/zivo/docs/readme.txt",
-    )
-    assert next_state.notification == NotificationState(
-        level="info",
-        message="Compressing as zip 2/5: readme.txt",
-    )
-
-def test_archive_extract_completed_requests_snapshot_for_destination_parent() -> None:
-    dest_parent = str(Path("/tmp/output").resolve())
-    dest_path = str(Path("/tmp/output/archive").resolve())
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_input=PendingInputState(
-            prompt="Extract to: ",
-            value=dest_path,
-            extract_source_path="/home/tadashi/develop/zivo/archive.zip",
-        ),
-        pending_archive_extract_request_id=9,
-    )
-
-    result = reduce_app_state(
-        state,
-        ArchiveExtractCompleted(
-            request_id=9,
-            result=ExtractArchiveResult(
-                destination_path=dest_path,
-                extracted_entries=2,
-                total_entries=2,
-                message="Extracted 2 entries to archive",
-            ),
-        ),
-    )
-
-    assert result.state.post_reload_notification == NotificationState(
-        level="info",
-        message="Extracted 2 entries to archive",
-    )
-    assert result.effects == (
-        LoadBrowserSnapshotEffect(
-            request_id=1,
-            path=dest_parent,
-            cursor_path=dest_path,
-            blocking=True,
-            invalidate_paths=browser_snapshot_invalidation_paths(
-                dest_parent, dest_path
-            ),
-        ),
-    )
-
-def test_zip_compress_completed_requests_snapshot_for_destination_parent() -> None:
-    dest_parent = str(Path("/tmp").resolve())
-    dest_path = str(Path("/tmp/output.zip").resolve())
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_input=PendingInputState(
-            prompt="Compress to: ",
-            value=dest_path,
-            zip_source_paths=("/home/tadashi/develop/zivo/docs",),
-        ),
-        pending_zip_compress_request_id=9,
-    )
-
-    result = reduce_app_state(
-        state,
-        ZipCompressCompleted(
-            request_id=9,
-            result=CreateZipArchiveResult(
-                destination_path=dest_path,
-                archived_entries=2,
-                total_entries=2,
-                message="Created output.zip with 2 entries",
-            ),
-        ),
-    )
-
-    assert result.state.post_reload_notification == NotificationState(
-        level="info",
-        message="Created output.zip with 2 entries",
-    )
-    assert result.effects == (
-        LoadBrowserSnapshotEffect(
-            request_id=1,
-            path=dest_parent,
-            cursor_path=dest_path,
-            blocking=True,
-            invalidate_paths=browser_snapshot_invalidation_paths(
-                dest_parent, dest_path
-            ),
-        ),
-    )
-
-def test_archive_extract_failed_returns_to_extract_mode() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_input=PendingInputState(
-            prompt="Extract to: ",
-            value="/tmp/output/archive",
-            extract_source_path="/home/tadashi/develop/zivo/archive.zip",
-        ),
-        pending_archive_extract_request_id=12,
-    )
-
-    next_state = _reduce_state(
-        state,
-        ArchiveExtractFailed(request_id=12, message="Unsupported archive member type: link"),
-    )
-
-    assert next_state.ui_mode == "EXTRACT"
-    assert next_state.pending_archive_extract_request_id is None
-    assert next_state.notification == NotificationState(
-        level="error",
-        message="Unsupported archive member type: link",
-    )
-
-def test_zip_compress_failed_returns_to_zip_mode() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_input=PendingInputState(
-            prompt="Compress to: ",
-            value="/tmp/output.zip",
-            zip_source_paths=("/home/tadashi/develop/zivo/docs",),
-        ),
-        pending_zip_compress_request_id=12,
-    )
-
-    next_state = _reduce_state(
-        state,
-        ZipCompressFailed(request_id=12, message="Destination path already exists as a directory"),
-    )
-
-    assert next_state.ui_mode == "ZIP"
-    assert next_state.pending_zip_compress_request_id is None
-    assert next_state.notification == NotificationState(
-        level="error",
-        message="Destination path already exists as a directory",
-    )
-
-def test_archive_preparation_failed_returns_to_extract_mode() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_input=PendingInputState(
-            prompt="Extract to: ",
-            value="/tmp/output/archive",
-            extract_source_path="/home/tadashi/develop/zivo/archive.zip",
-        ),
-        pending_archive_prepare_request_id=7,
-    )
-
-    next_state = _reduce_state(
-        state,
-        ArchivePreparationFailed(request_id=7, message="Unsupported archive format: archive.rar"),
-    )
-
-    assert next_state.ui_mode == "EXTRACT"
-    assert next_state.pending_archive_prepare_request_id is None
-    assert next_state.notification == NotificationState(
-        level="error",
-        message="Unsupported archive format: archive.rar",
-    )
-
-def test_set_pending_input_value_updates_current_value() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CREATE",
-        pending_input=PendingInputState(prompt="New file: ", value="", create_kind="file"),
-    )
-
-    next_state = _reduce_state(state, SetPendingInputValue("notes.txt", cursor_pos=8))
-
-    assert next_state.pending_input is not None
-    assert next_state.pending_input.value == "notes.txt"
-
-def test_submit_pending_input_rejects_duplicate_name() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CREATE",
-        pending_input=PendingInputState(
-            prompt="New file: ",
-            value="README.md",
-            create_kind="file",
-        ),
-    )
-
-    next_state = _reduce_state(state, SubmitPendingInput())
-
-    assert next_state.ui_mode == "CONFIRM"
-    assert next_state.pending_input is not None
-    assert next_state.notification is None
-    assert next_state.name_conflict == NameConflictState(
-        kind="create_file",
-        name="README.md",
-    )
-
-def test_submit_pending_input_treats_unchanged_rename_as_noop() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="RENAME",
-        pending_input=PendingInputState(
-            prompt="Rename: ",
-            value="docs",
-            target_path="/home/tadashi/develop/zivo/docs",
-        ),
-    )
-
-    next_state = _reduce_state(state, SubmitPendingInput())
-
-    assert next_state.ui_mode == "BROWSING"
-    assert next_state.pending_input is None
-    assert next_state.notification == NotificationState(level="info", message="Name unchanged")
-
-def test_submit_pending_input_emits_file_mutation_effect() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="RENAME",
-        pending_input=PendingInputState(
-            prompt="Rename: ",
-            value="manuals",
-            target_path="/home/tadashi/develop/zivo/docs",
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.state.ui_mode == "BUSY"
-    assert result.state.pending_file_mutation_request_id == 1
-    assert result.effects == (
-        RunFileMutationEffect(
-            request_id=1,
-            request=RenameRequest(
-                source_path="/home/tadashi/develop/zivo/docs",
-                new_name="manuals",
-            ),
-        ),
-    )
-
-def test_submit_pending_input_emits_create_effect() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CREATE",
-        pending_input=PendingInputState(
-            prompt="New file: ",
-            value="notes.txt",
-            create_kind="file",
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.effects == (
-        RunFileMutationEffect(
-            request_id=1,
-            request=CreatePathRequest(
-                parent_dir="/home/tadashi/develop/zivo",
-                name="notes.txt",
-                kind="file",
-            ),
-        ),
-    )
-
-def test_submit_pending_input_name_conflict_enters_confirm_mode_for_rename() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="RENAME",
-        pending_input=PendingInputState(
-            prompt="Rename: ",
-            value="src",
-            target_path="/home/tadashi/develop/zivo/docs",
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.state.ui_mode == "CONFIRM"
-    assert result.state.notification is None
-    assert result.state.name_conflict == NameConflictState(kind="rename", name="src")
-    assert result.effects == ()
-
-def test_submit_pending_input_name_conflict_enters_confirm_mode_for_create_dir() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CREATE",
-        pending_input=PendingInputState(
-            prompt="New directory: ",
-            value="docs",
-            create_kind="dir",
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.state.ui_mode == "CONFIRM"
-    assert result.state.name_conflict == NameConflictState(kind="create_dir", name="docs")
-    assert result.effects == ()
-
 def test_confirm_filter_input_returns_to_browsing() -> None:
     state = build_initial_app_state()
     state = _reduce_state(state, SetUiMode("FILTER"))
@@ -2356,617 +1387,6 @@ def test_cancel_filter_input_clears_query_from_browsing() -> None:
     assert next_state.ui_mode == "BROWSING"
     assert next_state.filter.query == ""
     assert next_state.filter.active is False
-
-def test_copy_targets_updates_clipboard_state() -> None:
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(state, CopyTargets(("/home/tadashi/develop/zivo/docs",)))
-
-    assert next_state.clipboard.mode == "copy"
-    assert next_state.clipboard.paths == ("/home/tadashi/develop/zivo/docs",)
-
-def test_copy_targets_warns_when_empty() -> None:
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(state, CopyTargets(()))
-
-    assert next_state.notification == NotificationState(level="warning", message="Nothing to copy")
-    assert next_state.clipboard.mode == "none"
-
-def test_cut_targets_warns_when_empty() -> None:
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(state, CutTargets(()))
-
-    assert next_state.notification == NotificationState(level="warning", message="Nothing to cut")
-    assert next_state.clipboard.mode == "none"
-
-def test_paste_clipboard_emits_paste_effect_and_sets_busy() -> None:
-    state = _reduce_state(
-        build_initial_app_state(),
-        CopyTargets(("/home/tadashi/develop/zivo/docs",)),
-    )
-
-    result = reduce_app_state(state, PasteClipboard())
-
-    assert result.state.ui_mode == "BUSY"
-    assert result.state.pending_paste_request_id == 1
-    assert result.effects == (
-        RunClipboardPasteEffect(
-            request_id=1,
-            request=result.effects[0].request,
-        ),
-    )
-    assert result.effects[0].request.destination_dir == "/home/tadashi/develop/zivo"
-
-def test_paste_clipboard_warns_when_empty() -> None:
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(state, PasteClipboard())
-
-    assert next_state.notification == NotificationState(
-        level="warning",
-        message="Clipboard is empty",
-    )
-    assert next_state.ui_mode == "BROWSING"
-
-def test_undo_last_operation_warns_when_stack_is_empty() -> None:
-    next_state = _reduce_state(build_initial_app_state(), UndoLastOperation())
-
-    assert next_state.notification == NotificationState(level="warning", message="Nothing to undo")
-
-def test_paste_needs_resolution_enters_confirm_mode() -> None:
-    state = _reduce_state(
-        build_initial_app_state(),
-        CopyTargets(("/home/tadashi/develop/zivo/docs",)),
-    )
-    requested = reduce_app_state(state, PasteClipboard()).state
-
-    conflict = PasteConflict(
-        source_path="/home/tadashi/develop/zivo/docs",
-        destination_path="/home/tadashi/develop/zivo/docs",
-    )
-    next_state = _reduce_state(
-        requested,
-        ClipboardPasteNeedsResolution(
-            request_id=1,
-            request=PasteRequest(
-                mode="copy",
-                source_paths=("/home/tadashi/develop/zivo/docs",),
-                destination_dir="/home/tadashi/develop/zivo",
-            ),
-            conflicts=(conflict,),
-        ),
-    )
-
-    assert next_state.ui_mode == "CONFIRM"
-    assert isinstance(next_state.paste_conflict, PasteConflictState)
-    assert next_state.paste_conflict.first_conflict == conflict
-
-def test_paste_needs_resolution_uses_configured_default_resolution() -> None:
-    state = _reduce_state(
-        build_initial_app_state(paste_conflict_action="rename"),
-        CopyTargets(("/home/tadashi/develop/zivo/docs",)),
-    )
-    requested = reduce_app_state(state, PasteClipboard()).state
-
-    conflict = PasteConflict(
-        source_path="/home/tadashi/develop/zivo/docs",
-        destination_path="/home/tadashi/develop/zivo/docs",
-    )
-    result = reduce_app_state(
-        requested,
-        ClipboardPasteNeedsResolution(
-            request_id=1,
-            request=PasteRequest(
-                mode="copy",
-                source_paths=("/home/tadashi/develop/zivo/docs",),
-                destination_dir="/home/tadashi/develop/zivo",
-            ),
-            conflicts=(conflict,),
-        ),
-    )
-
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects == (
-        RunClipboardPasteEffect(
-            request_id=2,
-            request=PasteRequest(
-                mode="copy",
-                source_paths=("/home/tadashi/develop/zivo/docs",),
-                destination_dir="/home/tadashi/develop/zivo",
-                conflict_resolution="rename",
-            ),
-        ),
-    )
-
-def test_paste_needs_resolution_ignores_stale_request() -> None:
-    state = _reduce_state(
-        build_initial_app_state(),
-        CopyTargets(("/home/tadashi/develop/zivo/docs",)),
-    )
-    requested = reduce_app_state(state, PasteClipboard()).state
-
-    conflict = PasteConflict(
-        source_path="/home/tadashi/develop/zivo/docs",
-        destination_path="/home/tadashi/develop/zivo/docs",
-    )
-    next_state = _reduce_state(
-        requested,
-        ClipboardPasteNeedsResolution(
-            request_id=99,
-            request=PasteRequest(
-                mode="copy",
-                source_paths=("/home/tadashi/develop/zivo/docs",),
-                destination_dir="/home/tadashi/develop/zivo",
-            ),
-            conflicts=(conflict,),
-        ),
-    )
-
-    assert next_state == requested
-
-def test_resolve_paste_conflict_restarts_paste_with_resolution() -> None:
-    conflict = PasteConflict(
-        source_path="/home/tadashi/develop/zivo/docs",
-        destination_path="/home/tadashi/develop/zivo/docs",
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        paste_conflict=PasteConflictState(
-            request=PasteRequest(
-                mode="copy",
-                source_paths=("/home/tadashi/develop/zivo/docs",),
-                destination_dir="/home/tadashi/develop/zivo",
-            ),
-            conflicts=(conflict,),
-            first_conflict=conflict,
-        ),
-        next_request_id=2,
-    )
-    state = replace(
-        state,
-        notification=None,
-    )
-
-    result = reduce_app_state(state, ResolvePasteConflict("rename"))
-
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects == (
-        RunClipboardPasteEffect(
-            request_id=2,
-            request=PasteRequest(
-                mode="copy",
-                source_paths=("/home/tadashi/develop/zivo/docs",),
-                destination_dir="/home/tadashi/develop/zivo",
-                conflict_resolution="rename",
-            ),
-        ),
-    )
-
-def test_clipboard_paste_completed_for_cut_clears_clipboard_and_requests_reload() -> None:
-    state = _reduce_state(
-        build_initial_app_state(),
-        CutTargets(("/home/tadashi/develop/zivo/docs",)),
-    )
-    state = replace(state, pending_paste_request_id=4)
-
-    result = reduce_app_state(
-        state,
-        ClipboardPasteCompleted(
-            request_id=4,
-            summary=PasteSummary(
-                mode="cut",
-                destination_dir="/home/tadashi/develop/zivo",
-                total_count=1,
-                success_count=1,
-                skipped_count=0,
-            ),
-            applied_changes=(
-                PasteAppliedChange(
-                    source_path="/tmp/staging/docs",
-                    destination_path="/home/tadashi/develop/zivo/docs",
-                ),
-            ),
-        ),
-    )
-
-    assert result.state.clipboard.mode == "none"
-    assert result.state.undo_stack == (
-        UndoEntry(
-            kind="paste_cut",
-            steps=(
-                UndoMovePathStep(
-                    source_path="/home/tadashi/develop/zivo/docs",
-                    destination_path="/tmp/staging/docs",
-                ),
-            ),
-        ),
-    )
-    assert result.state.pending_browser_snapshot_request_id == 1
-    assert result.effects == (
-        LoadBrowserSnapshotEffect(
-            request_id=1,
-            path="/home/tadashi/develop/zivo",
-            cursor_path="/home/tadashi/develop/zivo/docs",
-            blocking=False,
-            invalidate_paths=tuple(
-                str(Path(p).resolve())
-                for p in (
-                    "/home/tadashi/develop/zivo",
-                    "/home/tadashi/develop",
-                    "/home/tadashi/develop/zivo/docs",
-                )
-            ),
-        ),
-    )
-
-def test_clipboard_paste_completed_pushes_copy_undo_entry() -> None:
-    state = replace(build_initial_app_state(), pending_paste_request_id=4)
-
-    next_state = _reduce_state(
-        state,
-        ClipboardPasteCompleted(
-            request_id=4,
-            summary=PasteSummary(
-                mode="copy",
-                destination_dir="/home/tadashi/develop/zivo",
-                total_count=1,
-                success_count=1,
-                skipped_count=0,
-            ),
-            applied_changes=(
-                PasteAppliedChange(
-                    source_path="/tmp/source/docs",
-                    destination_path="/home/tadashi/develop/zivo/docs copy",
-                ),
-            ),
-        ),
-    )
-
-    assert next_state.undo_stack == (
-        UndoEntry(
-            kind="paste_copy",
-            steps=(UndoDeletePathStep(path="/home/tadashi/develop/zivo/docs copy"),),
-        ),
-    )
-
-def test_clipboard_paste_completed_skips_undo_for_overwrite() -> None:
-    state = replace(build_initial_app_state(), pending_paste_request_id=4)
-
-    next_state = _reduce_state(
-        state,
-        ClipboardPasteCompleted(
-            request_id=4,
-            summary=PasteSummary(
-                mode="copy",
-                destination_dir="/home/tadashi/develop/zivo",
-                total_count=1,
-                success_count=1,
-                skipped_count=0,
-                overwrote_count=1,
-            ),
-            applied_changes=(
-                PasteAppliedChange(
-                    source_path="/tmp/source/docs",
-                    destination_path="/home/tadashi/develop/zivo/docs",
-                ),
-            ),
-        ),
-    )
-
-    assert next_state.undo_stack == ()
-    assert next_state.post_reload_notification == NotificationState(
-        level="info",
-        message="Copied 1 item(s), undo unavailable for overwritten items",
-    )
-
-def test_file_mutation_completed_requests_reload_with_result_cursor() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_file_mutation_request_id=4,
-        pending_input=PendingInputState(
-            prompt="New file: ",
-            value="notes.txt",
-            create_kind="file",
-        ),
-    )
-
-    result = reduce_app_state(
-        state,
-        FileMutationCompleted(
-            request_id=4,
-            result=FileMutationResult(
-                path="/home/tadashi/develop/zivo/notes.txt",
-                message="Created file notes.txt",
-            ),
-        ),
-    )
-
-    assert result.state.ui_mode == "BROWSING"
-    assert result.state.pending_input is None
-    assert result.state.undo_stack == ()
-    assert result.state.pending_browser_snapshot_request_id == 1
-    assert result.effects == (
-        LoadBrowserSnapshotEffect(
-            request_id=1,
-            path="/home/tadashi/develop/zivo",
-            cursor_path="/home/tadashi/develop/zivo/notes.txt",
-            blocking=False,
-            invalidate_paths=tuple(
-                str(Path(p).resolve())
-                for p in (
-                    "/home/tadashi/develop/zivo",
-                    "/home/tadashi/develop",
-                    "/home/tadashi/develop/zivo/notes.txt",
-                )
-            ),
-        ),
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_file_mutation_request_id=4,
-        pending_input=PendingInputState(
-            prompt="Rename: ",
-            value="manuals",
-            target_path="/home/tadashi/develop/zivo/docs",
-        ),
-    )
-
-    next_state = _reduce_state(
-        state,
-        FileMutationCompleted(
-            request_id=4,
-            result=FileMutationResult(
-                path="/home/tadashi/develop/zivo/manuals",
-                message="Renamed to manuals",
-                operation="rename",
-                source_path="/home/tadashi/develop/zivo/docs",
-            ),
-        ),
-    )
-
-    assert next_state.undo_stack == (
-        UndoEntry(
-            kind="rename",
-            steps=(
-                UndoMovePathStep(
-                    source_path="/home/tadashi/develop/zivo/manuals",
-                    destination_path="/home/tadashi/develop/zivo/docs",
-                ),
-            ),
-        ),
-    )
-
-def test_delete_file_mutation_completed_requests_reload_without_deleted_cursor() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_file_mutation_request_id=7,
-        current_pane=replace(
-            build_initial_app_state().current_pane,
-            cursor_path="/home/tadashi/develop/zivo/docs",
-        ),
-    )
-
-    result = reduce_app_state(
-        state,
-        FileMutationCompleted(
-            request_id=7,
-            result=FileMutationResult(
-                path=None,
-                message="Trashed 1 item",
-                removed_paths=("/home/tadashi/develop/zivo/docs",),
-            ),
-        ),
-    )
-
-    assert result.state.ui_mode == "BROWSING"
-    assert result.state.undo_stack == ()
-    assert result.effects == (
-        LoadBrowserSnapshotEffect(
-            request_id=1,
-            path="/home/tadashi/develop/zivo",
-            cursor_path="/home/tadashi/develop/zivo/src",
-            blocking=False,
-            invalidate_paths=tuple(
-                str(Path(p).resolve())
-                for p in (
-                    "/home/tadashi/develop/zivo",
-                    "/home/tadashi/develop",
-                    "/home/tadashi/develop/zivo/src",
-                )
-            ),
-        ),
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_file_mutation_request_id=7,
-    )
-
-    next_state = _reduce_state(
-        state,
-        FileMutationCompleted(
-            request_id=7,
-            result=FileMutationResult(
-                path=None,
-                message="Trashed 1 item",
-                removed_paths=("/home/tadashi/develop/zivo/docs",),
-                operation="delete",
-                delete_mode="trash",
-                trash_records=(
-                    TrashRestoreRecord(
-                        original_path="/home/tadashi/develop/zivo/docs",
-                        trashed_path="/home/tadashi/.local/share/Trash/files/docs",
-                        metadata_path="/home/tadashi/.local/share/Trash/info/docs.trashinfo",
-                    ),
-                ),
-            ),
-        ),
-    )
-
-    assert next_state.undo_stack == (
-        UndoEntry(
-            kind="trash_delete",
-            steps=(
-                UndoRestoreTrashStep(
-                    record=TrashRestoreRecord(
-                        original_path="/home/tadashi/develop/zivo/docs",
-                        trashed_path="/home/tadashi/.local/share/Trash/files/docs",
-                        metadata_path="/home/tadashi/.local/share/Trash/info/docs.trashinfo",
-                    )
-                ),
-            ),
-        ),
-    )
-
-def test_undo_last_operation_runs_effect() -> None:
-    entry = UndoEntry(kind="paste_copy", steps=(UndoDeletePathStep(path="/tmp/copied"),))
-    state = replace(build_initial_app_state(), undo_stack=(entry,), next_request_id=6)
-
-    result = reduce_app_state(state, UndoLastOperation())
-
-    assert result.state.pending_undo_entry == entry
-    assert result.state.pending_undo_request_id == 6
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects == (RunUndoEffect(request_id=6, entry=entry),)
-
-def test_undo_completed_pops_stack_and_requests_reload() -> None:
-    entry = UndoEntry(kind="paste_copy", steps=(UndoDeletePathStep(path="/tmp/copied"),))
-    state = replace(
-        build_initial_app_state(),
-        undo_stack=(entry,),
-        pending_undo_entry=entry,
-        pending_undo_request_id=9,
-        next_request_id=4,
-    )
-
-    result = reduce_app_state(
-        state,
-        UndoCompleted(
-            request_id=9,
-            entry=entry,
-            result=UndoResult(
-                path=None,
-                message="Undid copied item",
-                removed_paths=("/tmp/copied",),
-            ),
-        ),
-    )
-
-    assert result.state.undo_stack == ()
-    assert result.state.pending_undo_request_id is None
-    assert result.state.post_reload_notification == NotificationState(
-        level="info",
-        message="Undid copied item",
-    )
-    assert result.effects == (
-        LoadBrowserSnapshotEffect(
-            request_id=4,
-            path="/home/tadashi/develop/zivo",
-            cursor_path="/home/tadashi/develop/zivo/docs",
-            blocking=False,
-            invalidate_paths=tuple(
-                str(Path(p).resolve())
-                for p in (
-                    "/home/tadashi/develop/zivo",
-                    "/home/tadashi/develop",
-                    "/home/tadashi/develop/zivo/docs",
-                )
-            ),
-        ),
-    )
-    entry = UndoEntry(kind="paste_copy", steps=(UndoDeletePathStep(path="/tmp/copied"),))
-    state = replace(
-        build_initial_app_state(),
-        undo_stack=(entry,),
-        pending_undo_entry=entry,
-        pending_undo_request_id=9,
-        ui_mode="BUSY",
-    )
-
-    next_state = _reduce_state(state, UndoFailed(request_id=9, message="permission denied"))
-
-    assert next_state.pending_undo_request_id is None
-    assert next_state.undo_stack == (entry,)
-    assert next_state.notification == NotificationState(
-        level="error",
-        message="permission denied",
-    )
-
-def test_file_mutation_failed_keeps_input_value_and_returns_error() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_file_mutation_request_id=3,
-        pending_input=PendingInputState(
-            prompt="Rename: ",
-            value="docs copy",
-            target_path="/home/tadashi/develop/zivo/docs",
-        ),
-    )
-
-    next_state = _reduce_state(state, FileMutationFailed(request_id=3, message="permission denied"))
-
-    assert next_state.ui_mode == "RENAME"
-    assert next_state.pending_input is not None
-    assert next_state.pending_input.value == "docs copy"
-    assert next_state.notification == NotificationState(
-        level="error",
-        message="permission denied",
-    )
-
-def test_delete_file_mutation_failed_returns_to_browsing() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_file_mutation_request_id=5,
-    )
-
-    next_state = _reduce_state(state, FileMutationFailed(request_id=5, message="trash failed"))
-
-    assert next_state.ui_mode == "BROWSING"
-    assert next_state.notification == NotificationState(
-        level="error",
-        message="trash failed",
-    )
-
-def test_clipboard_paste_failed_returns_to_browsing_and_clears_dialog_state() -> None:
-    conflict = PasteConflict(
-        source_path="/home/tadashi/develop/zivo/docs",
-        destination_path="/home/tadashi/develop/zivo/docs",
-    )
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="BUSY",
-        pending_paste_request_id=4,
-        paste_conflict=PasteConflictState(
-            request=PasteRequest(
-                mode="copy",
-                source_paths=("/home/tadashi/develop/zivo/docs",),
-                destination_dir="/home/tadashi/develop/zivo",
-            ),
-            conflicts=(conflict,),
-            first_conflict=conflict,
-        ),
-        delete_confirmation=DeleteConfirmationState(paths=("/home/tadashi/develop/zivo/docs",)),
-        name_conflict=NameConflictState(kind="rename", name="docs"),
-    )
-
-    next_state = _reduce_state(
-        state,
-        ClipboardPasteFailed(request_id=4, message="paste failed"),
-    )
-
-    assert next_state.ui_mode == "BROWSING"
-    assert next_state.paste_conflict is None
-    assert next_state.delete_confirmation is None
-    assert next_state.name_conflict is None
-    assert next_state.notification == NotificationState(level="error", message="paste failed")
 
 def test_external_launch_failed_sets_error_notification() -> None:
     state = build_initial_app_state()
@@ -3039,49 +1459,6 @@ def test_dismiss_name_conflict_restores_create_mode_and_keeps_input() -> None:
     assert next_state.ui_mode == "CREATE"
     assert next_state.pending_input == state.pending_input
     assert next_state.name_conflict is None
-
-def test_cancel_paste_conflict_returns_to_browsing_with_warning() -> None:
-    state = replace(build_initial_app_state(), ui_mode="CONFIRM")
-
-    next_state = _reduce_state(state, CancelPasteConflict())
-
-    assert next_state.ui_mode == "BROWSING"
-    assert next_state.notification == NotificationState(level="warning", message="Paste cancelled")
-
-def test_toggle_selection_and_advance_moves_cursor_to_next_visible_entry() -> None:
-    state = build_initial_app_state()
-    current_path = "/home/tadashi/develop/zivo/docs"
-    visible_paths = (
-        "/home/tadashi/develop/zivo/docs",
-        "/home/tadashi/develop/zivo/src",
-        "/home/tadashi/develop/zivo/tests",
-        "/home/tadashi/develop/zivo/README.md",
-        "/home/tadashi/develop/zivo/pyproject.toml",
-    )
-
-    result = reduce_app_state(
-        state,
-        ToggleSelectionAndAdvance(path=current_path, visible_paths=visible_paths),
-    )
-
-    assert result.state.current_pane.selected_paths == frozenset({current_path})
-    assert result.state.current_pane.cursor_path == "/home/tadashi/develop/zivo/src"
-    assert result.state.pending_child_pane_request_id == 1
-    assert result.effects == (
-        LoadChildPaneSnapshotEffect(
-            request_id=1,
-            current_path="/home/tadashi/develop/zivo",
-            cursor_path="/home/tadashi/develop/zivo/src",
-        ),
-        RunDirectorySizeEffect(
-            request_id=2,
-            paths=(
-                "/home/tadashi/develop/zivo/docs",
-                "/home/tadashi/develop/zivo/src",
-                "/home/tadashi/develop/zivo/tests",
-            ),
-        ),
-    )
 
 def test_move_cursor_and_select_range_sets_anchor_and_selects_contiguous_entries() -> None:
     state = build_initial_app_state()
@@ -4285,32 +2662,6 @@ def test_browser_snapshot_loaded_updates_inactive_tab_only() -> None:
     loaded = _reduce_state(loaded, ActivateNextTab())
     assert loaded.current_path == "/tmp/project"
 
-def test_begin_empty_trash_enters_confirm_mode(monkeypatch) -> None:
-    monkeypatch.setattr("platform.system", lambda: "Darwin")
-
-    state = build_initial_app_state()
-
-    next_state = _reduce_state(state, BeginEmptyTrash())
-
-    assert next_state.ui_mode == "CONFIRM"
-    assert next_state.empty_trash_confirmation is not None
-    assert next_state.empty_trash_confirmation.platform == "darwin"
-    assert next_state.command_palette is None
-    assert next_state.pending_input is None
-
-def test_cancel_empty_trash_confirmation_returns_to_browsing() -> None:
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        empty_trash_confirmation=EmptyTrashConfirmationState(platform="darwin"),
-    )
-
-    next_state = _reduce_state(state, CancelEmptyTrashConfirmation())
-
-    assert next_state.ui_mode == "BROWSING"
-    assert next_state.empty_trash_confirmation is None
-    assert next_state.notification is None
-
 def test_set_pending_key_sequence_updates_state() -> None:
     state = build_initial_app_state()
 
@@ -4324,153 +2675,6 @@ def test_set_pending_key_sequence_updates_state() -> None:
         possible_next_keys=("y",),
     )
 
-def test_confirm_empty_trash_shows_notification_on_success(monkeypatch) -> None:
-    from unittest.mock import MagicMock
-
-    from zivo.services.trash_operations import MacOsTrashService
-
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CONFIRM",
-        empty_trash_confirmation=EmptyTrashConfirmationState(platform="darwin"),
-    )
-
-    fake_service = MacOsTrashService()
-    monkeypatch.setattr(
-        "zivo.services.resolve_trash_service",
-        lambda: fake_service,
-    )
-    monkeypatch.setattr(
-        "zivo.services.trash_operations.subprocess.run",
-        lambda *a, **kw: MagicMock(returncode=0, stderr=""),
-    )
-
-    class FakeHome:
-        def __truediv__(self, other):
-            return FakePath()
-
-    class FakePath:
-        def exists(self):
-            return False
-
-    monkeypatch.setattr("pathlib.Path.home", lambda: FakeHome())
-
-    next_state = _reduce_state(state, ConfirmEmptyTrash())
-
-    assert next_state.ui_mode == "BROWSING"
-    assert next_state.empty_trash_confirmation is None
-    assert next_state.notification is not None
-    assert next_state.notification.level == "info"
-
-def test_submit_pending_input_case_only_rename_emits_mutation() -> None:
-    """Case-only rename (docs -> Docs) should proceed, not show 'Name unchanged'."""
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="RENAME",
-        pending_input=PendingInputState(
-            prompt="Rename: ",
-            value="Docs",
-            target_path="/home/tadashi/develop/zivo/docs",
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects == (
-        RunFileMutationEffect(
-            request_id=1,
-            request=RenameRequest(
-                source_path="/home/tadashi/develop/zivo/docs",
-                new_name="Docs",
-            ),
-        ),
-    )
-
-def test_submit_pending_input_case_insensitive_conflict_on_macos(monkeypatch) -> None:
-    """On macOS, renaming to a case-different existing name should conflict."""
-    import zivo.state.reducer_common as rc
-
-    monkeypatch.setattr(rc, "_is_macos", True)
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="RENAME",
-        pending_input=PendingInputState(
-            prompt="Rename: ",
-            value="Src",
-            target_path="/home/tadashi/develop/zivo/docs",
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.state.ui_mode == "CONFIRM"
-    assert result.state.name_conflict == NameConflictState(kind="rename", name="Src")
-
-def test_submit_pending_input_case_sensitive_no_conflict_on_linux(monkeypatch) -> None:
-    """On Linux, renaming to a case-different existing name should NOT conflict."""
-    import zivo.state.reducer_common as rc
-
-    monkeypatch.setattr(rc, "_is_macos", False)
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="RENAME",
-        pending_input=PendingInputState(
-            prompt="Rename: ",
-            value="Src",
-            target_path="/home/tadashi/develop/zivo/docs",
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects != ()
-
-def test_submit_pending_input_rejects_colon_in_name_on_macos(monkeypatch) -> None:
-    """On macOS, names containing ':' should be rejected."""
-    import zivo.state.reducer_common as rc
-
-    monkeypatch.setattr(rc, "_is_macos", True)
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CREATE",
-        pending_input=PendingInputState(
-            prompt="New file: ",
-            value="file:name.txt",
-            create_kind="file",
-        ),
-    )
-
-    next_state = _reduce_state(state, SubmitPendingInput())
-
-    assert next_state.notification is not None
-    assert next_state.notification.level == "error"
-    assert "colons" in next_state.notification.message
-
-def test_submit_pending_input_allows_colon_in_name_on_linux(monkeypatch) -> None:
-    """On Linux, names containing ':' should not be rejected."""
-    import zivo.state.reducer_common as rc
-
-    monkeypatch.setattr(rc, "_is_macos", False)
-    state = replace(
-        build_initial_app_state(),
-        ui_mode="CREATE",
-        pending_input=PendingInputState(
-            prompt="New file: ",
-            value="file:name.txt",
-            create_kind="file",
-        ),
-    )
-
-    result = reduce_app_state(state, SubmitPendingInput())
-
-    # Should proceed to emit a create effect (not blocked by colon validation)
-    assert result.state.ui_mode == "BUSY"
-    assert result.effects != ()
-
-
 # ---------------------------------------------------------------------------
 # Find-and-replace (replace_in_found_files) tests
 # ---------------------------------------------------------------------------
-
